requirements.txt
django>=1.8
django-registration-redux==1.6
django-phonenumber-field==1.1
djangorestframework==3.6.3